Evolve: cell with 3 neighbours come alive
This was when I realized that I had a bug (I unskipped and I saw I had
something failing) and started running through the cases and found the
failing one.

I realized I birthed cells when there were only 2 cells nearby, which
is not okay. So refactor to something that works and here we go.